2024-python/requirements.txt
nographs==3.4.0
numpy==2.1.3
PyYAML==6.0.2
termcolor==2.5.0
tqdm==4.67.1
